+++ flow/src/main/java/flow/Flow.java
@@ -24,17 +24,31 @@ public final class Flow {
     FORWARD, BACKWARD, REPLACE
   }
 
+  /** Supplied by Flow to the Listener, which is responsible for calling onComplete(). */
+  public interface Callback {
+    /**
+     * Must be called exactly once to indicate that the corresponding transition has completed.
+     *
+     * If not called, the backstack will not be updated and further calls to Flow will not execute.
+     * Calling more than once will result in an exception.
+     */
+    void onComplete();
+  }
+
   public interface Listener {
     /**
-     * Notifies the listener that the backstack is about to change. Note that the change
-     * will not take effect until this method returns. That is, {@code nextBackstack} is where
-     * the Flow is going next, and {@link Flow#getBackstack()} is where it's coming from.
+     * Notifies the listener that the backstack is about to change. Note that the backstack of
+     * is not actually changed until the callback is triggered.  That is, {@code nextBackstack} is
+     * where the Flow is going next, and {@link Flow#getBackstack()} is where it's coming from.
+     *
+     * @param callback Must be called to indicate completion.
      */
-    void go(Backstack nextBackstack, Direction direction);
+    void go(Backstack nextBackstack, Direction direction, Callback callback);
   }
 
   private final Listener listener;
   private Backstack backstack;
+  private Transition transition;
 
   public Flow(Backstack backstack, Listener listener) {
     this.listener = listener;
@@ -46,48 +60,65 @@ public final class Flow {
   }
 
   /** Push the screen onto the backstack. */
-  public void goTo(Object screen) {
-    Backstack newBackstack = backstack.buildUpon().push(screen).build();
-    forward(newBackstack);
+  public void goTo(final Object screen) {
+    move(new Transition() {
+      @Override public void execute() {
+        Backstack newBackstack = backstack.buildUpon().push(screen).build();
+        go(newBackstack, Direction.FORWARD);
+      }
+    });
   }
 
   /**
    * Reset to the specified screen. Pops until the screen is found.  If the screen is not found, the
    * entire backstack is replaced with the screen.
    */
-  public void resetTo(Object screen) {
-    Backstack.Builder builder = backstack.buildUpon();
-    int count = 0;
-    // Take care to leave the original screen instance on the stack, if we find it.  This enables
-    // some arguably bad behavior on the part of clients, but it's still probably the right thing
-    // to do.
-    Object lastPopped = null;
-    for (Iterator<Backstack.Entry> it = backstack.reverseIterator(); it.hasNext();) {
-      Backstack.Entry entry = it.next();
+  public void resetTo(final Object screen) {
+    move(new Transition() {
+      @Override public void execute() {
+        Backstack.Builder builder = backstack.buildUpon();
+        int count = 0;
+        // Take care to leave the original screen instance on the stack, if we find it.  This enables
+        // some arguably bad behavior on the part of clients, but it's still probably the right thing
+        // to do.
+        Object lastPopped = null;
+        for (Iterator<Backstack.Entry> it = backstack.reverseIterator(); it.hasNext();) {
+          Backstack.Entry entry = it.next();
 
-      if (entry.getScreen().equals(screen)) {
-        // Clear up to the target screen.
-        for (int i = 0; i < backstack.size() - count; i++) {
-          lastPopped = builder.pop().getScreen();
+          if (entry.getScreen().equals(screen)) {
+            // Clear up to the target screen.
+            for (int i = 0; i < backstack.size() - count; i++) {
+              lastPopped = builder.pop().getScreen();
+            }
+            break;
+          } else {
+            count++;
+          }
         }
-        break;
-      } else {
-        count++;
-      }
-    }
 
-    if (lastPopped != null) {
-      builder.push(lastPopped);
-      backward(builder.build());
-    } else {
-      builder.push(screen);
-      forward(builder.build());
-    }
+        Backstack newBackstack;
+        if (lastPopped != null) {
+          builder.push(lastPopped);
+          newBackstack = builder.build();
+          go(newBackstack, Direction.BACKWARD);
+        } else {
+          builder.push(screen);
+          newBackstack = builder.build();
+          go(newBackstack, Direction.FORWARD);
+        }
+      }
+    });
+
   }
 
   /** Replaces the current backstack with the up stack of the screen. */
-  public void replaceTo(Object screen) {
-    replace(preserveEquivalentPrefix(backstack, Backstack.fromUpChain(screen)));
+  public void replaceTo(final Object screen) {
+    move(new Transition() {
+      @Override public void execute() {
+        Backstack newBackstack = preserveEquivalentPrefix(backstack, Backstack.fromUpChain(screen));
+        go(newBackstack, Direction.REPLACE);
+      }
+    });
   }
 
   /**
@@ -95,14 +126,24 @@ public final class Flow {
    * @return false if going up is not possible.
    */
   public boolean goUp() {
-    Object current = backstack.current().getScreen();
-    if (current instanceof HasParent<?>) {
-      Object parent = ((HasParent) current).getParent();
-      backward(preserveEquivalentPrefix(backstack, Backstack.fromUpChain(parent)));
-      return true;
-    } else {
-      return false;
+    boolean canGoUp = false;
+    if (backstack.current().getScreen() instanceof HasParent || (transition != null && !transition.finished)) {
+      canGoUp = true;
     }
+    move(new Transition() {
+      @Override public void execute() {
+        Object current = backstack.current().getScreen();
+        if (current instanceof HasParent<?>) {
+          Object parent = ((HasParent) current).getParent();
+          Backstack newBackstack = preserveEquivalentPrefix(backstack, Backstack.fromUpChain(parent));
+          go(newBackstack, Direction.BACKWARD);
+        } else {
+          // We are not calling the listener, so we must complete this noop transition ourselves.
+          onComplete();
+        }
+      }
+    });
+    return canGoUp;
   }
 
   /**
@@ -110,33 +151,49 @@ public final class Flow {
    * @return false if going back is not possible.
    */
   public boolean goBack() {
-    if (backstack.size() == 1) {
-      return false;
-    }
+    boolean canGoBack = backstack.size() > 1 || (transition != null && !transition.finished);
+    move(new Transition() {
+      @Override public void execute() {
+        if (backstack.size() == 1) {
+          // We are not calling the listener, so we must complete this noop transition ourselves.
+          onComplete();
+        } else {
+          Backstack.Builder builder = backstack.buildUpon();
+          builder.pop();
+          Backstack newBackstack = builder.build();
+          go(newBackstack, Direction.BACKWARD);
+        }
+      }
+    });
 
-    Backstack.Builder builder = backstack.buildUpon();
-    builder.pop();
-    backward(builder.build());
-
-    return true;
+    return canGoBack;
   }
 
   /** Goes forward to a new backstack. */
-  public void forward(Backstack newBackstack) {
-    listener.go(newBackstack, Direction.FORWARD);
-    backstack = newBackstack;
+  public void forward(final Backstack newBackstack) {
+    move(new Transition() {
+      @Override public void execute() {
+        go(newBackstack, Direction.FORWARD);
+      }
+    });
   }
 
   /** Goes backward to a new backstack. */
-  public void backward(Backstack newBackstack) {
-    listener.go(newBackstack, Direction.BACKWARD);
-    backstack = newBackstack;
+  public void backward(final Backstack newBackstack) {
+    move(new Transition() {
+      @Override public void execute() {
+        go(newBackstack, Direction.BACKWARD);
+      }
+    });
   }
 
-  /** Replaces to a new backstack. */
-  private void replace(Backstack newBackstack) {
-    listener.go(newBackstack, Direction.REPLACE);
-    backstack = newBackstack;
+  private void move(Transition transition) {
+    if (this.transition == null || this.transition.finished) {
+      this.transition = transition;
+      transition.execute();
+    } else {
+      this.transition.enqueue(transition);
+    }
   }
 
   private static Backstack preserveEquivalentPrefix(Backstack current, Backstack proposed) {
@@ -164,5 +221,40 @@ public final class Flow {
       preserving.push(newIt.next().getScreen());
     }
     return preserving.build();
+  }
+
+  private abstract class Transition implements Callback {
+    boolean finished;
+    Transition next;
+    Backstack nextBackstack;
+
+    void enqueue(Transition transition) {
+      if (this.next == null) {
+        this.next = transition;
+      } else {
+        this.next.enqueue(transition);
+      }
+    }
+
+    @Override public void onComplete() {
+      if (finished) {
+        throw new IllegalStateException("onComplete already called for this transition");
+      }
+      if (nextBackstack != null) {
+        backstack = nextBackstack;
+      }
+      finished = true;
+      if (next != null) {
+        transition = next;
+        transition.execute();
+      }
+    }
+
+    protected void go(Backstack nextBackstack, Direction direction) {
+      this.nextBackstack = nextBackstack;
+      listener.go(nextBackstack, direction, this);
+    }
+
+    protected abstract void execute();
   }
 }